requirements.txt
Django>=4.0
tensorflow
pillow
numpy
requests
flask